Always return unicode when encoding is given (closes #18)
As reported in github issue #18 when a file-like object is
given and the encoding param is passed configobj fails to
decode the data on Python 2. This is a regression from
the 4.7 releases.

Fix is simply to check first for binary types in _decode()
helper as six.string_types includes the bytes str type
on Python 2.