Cleanup unrepr function and related code
There was some cruft left around from the old implemention
of unrepr, which can now be deleted.

Not sure if the lazy import of ast is worthwhile but the
implemention of that is now pretty harmless.

Note this does remove the exported but unused, and only
ever internally propogated, UnknownType exception type.